+++ src/pages/Login/components/Login/LoginComponent.tsx
@@ -1,22 +1,14 @@
-import { FormEvent, useState } from "react"
+import { FormEvent } from "react"
 import * as S from "./LoginComponent.Styles"
 import { theme } from "@/styles/theme"
-
 import LoginInputContainer from "./LoginInput/LoginInputContainer"
-import { API } from "@/apis/Api"
-import useAuthUserStore from "@/stores/useAuthUserStore"
+import useLogin from "../../hooks/useLogin"
 import { useNavigate } from "react-router-dom"
-import useModal from "@/components/Modal/hooks/useModal"
-import AlertModal from "@/components/Modal/components/AlertModal/AlertModal"
-import useLogin from "../hooks/useLogin"
 
 const LoginComponent = () => {
-  const { userInfoRef, updateUserInfo } = useLogin()
-  const [alertMessage, setAlertMessage] = useState("")
-
-  const { setLogin } = useAuthUserStore()
+  const { userInfoRef, updateUserInfo, AlertModalComponent, LoginApi_send } =
+    useLogin()
   const navigate = useNavigate()
-  const { isShowModal, showModal, closeModal } = useModal()
 
   const handleLogin = (event: FormEvent<HTMLFormElement>) => {
     event.preventDefault()
@@ -25,29 +17,12 @@ const LoginComponent = () => {
     if (email === "" || password === "") {
       return
     }
-
-    API.post("/login", {
-      email,
-      password,
-    })
-      .then((res) => {
-        const { user, token } = res.data
-        setLogin(user, token)
-        navigate("/", { replace: true })
-      })
-      .catch(() => {
-        setAlertMessage("잘못된 이메일이거나 잘못된 비밀번호의 조합입니다.")
-        showModal()
-      })
+    LoginApi_send.mutate({ email, password })
   }
 
   return (
     <>
-      <AlertModal
-        isShow={isShowModal}
-        alertMessage={alertMessage}
-        onClose={closeModal}
-      />
+      {AlertModalComponent}
       <S.LoginComponentLayout>
         <S.LoginComponentTitle>로그인</S.LoginComponentTitle>
         <S.LoginForm onSubmit={handleLogin}>
